Version 1. Basic script structure
This script will download a playlist from YouTube and store it in the specified directory. For now, it only works on playlists, not on individual videos. Furthermore, it is done linearly in terms of downloads. Just enter the URL of the  playlist, and a directory path, and it will be saved to the specified location.